requirements.txt
discord.py==2.6.4
cryptography==41.0.0
Pillow
